refactor: change import/export sequence, rename classes files
