+++ src/components/common/Component.ts
@@ -1,7 +1,7 @@
 abstract class Component {
   protected container: HTMLElement;
 
-  constructor(tagName: string, className: string) {
+  protected constructor(tagName: string, className: string) {
     this.container = document.createElement(tagName);
     this.container.className = className;
   }
