+++ src/components/common/Page.ts
@@ -2,7 +2,7 @@ abstract class Page {
   protected container: HTMLElement;
   static TextObject = {};
 
-  constructor(id: string) {
+  protected constructor(id: string) {
     this.container = document.createElement('main');
     this.container.id = id;
   }
