+++ src/components/pages/main/products/controller/controller.ts
@@ -1,5 +1,6 @@
-import AppLoader from './appLoader';
 import { ProductType } from '../../../../types';
+
+import AppLoader from './appLoader';
 
 class AppController extends AppLoader {
   getSources(callback: (data?: Array<ProductType>) => void) {
